refactor(core): de-monolith Stage 3 — relocate the FROZEN kinematics-contract talisman
Moves the single most safety-critical file — the deterministic vehicle flight-envelope
contract (`EnforceAction` / `DenyCode` / `VehicleKinematicsContract` /
`validate_vehicle_command`, 1326 lines) — VERBATIM from `src/gateway/kinematics_contract.rs`
to `kirra-core`, so the certified governor surface no longer lives in the heavy
verifier-service crate.

Safe-to-move by construction: the talisman imports ONLY `serde` and has ZERO
`crate::`/`super::` references (except a test's `use super::*`) — fully self-contained, so
the relocation needed no dependency rewiring. Content verified byte-identical (diff of the
two files, modulo the top path comment).

  - `crates/kirra-core/src/kinematics_contract.rs` — the talisman, verbatim (+ its 42 unit
    tests, which now run in kirra-core).
  - `src/gateway/kinematics_contract.rs` — now `pub use kirra_core::kinematics_contract::*;`
    (no `pub(crate)` items, so the `::*` re-export is complete). Every existing
    `…gateway::kinematics_contract::*` path holds, same type — the 10+ dependents (adapter,
    fabric router/governor, wcet_gate, verifier_store, audit_writer, the other gateway
    modules, the proptests) are untouched.

The talisman rule is honored: the contract is NEVER modified, only relocated.

Gate (all green): kirra-core 42 talisman tests pass + stays lean (0 heavy deps via
`cargo tree`); root crate 591 lib tests pass (633 − 42 that moved) incl. the gateway
`kinematics_proptest` via the re-export; adapter contract tests pass (the byte-identical
`validate_vehicle_command` path); parko-kirra tests pass (bridges to the contract);
clippy clean. The ISO-26262 invariant + WCET CI jobs are the final confirmation.
